Add a directive to build an ES6 version for FFx
Moved KintoClient logic to base class.

So we can derive specific target versions:

- index.js contains the main Web oriented client
- fx-src/index.js contains the Firefox integration version
